+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * `flexberry-groupedit` component:
     * Now possible specify `developerUserSettings` for `flexberry-groupedit` on `edit-form`.
 * `log` service: settings for displaying exceptions in the application log.
+* Add asyng loading record for `object-list-view` component.
 
 ### Changed
 * Updated dependency on `ember-flexberry-data` addon to v0.6.2-beta.4 version.